Update `findAllByUserId` and `deleteTodoById` to handle `userId` as `string | undefined`
* **findAllByUserId**
  - Change argument `userId` to `string | undefined`
  - Add check for `userId` and throw error if `userId` is `undefined`

* **deleteTodoById**
  - Change argument `userId` to `string | undefined`
  - Add check for `userId` and throw error if `userId` is `undefined`

+++ src/repositories/todo_repository.ts
@@ -6,10 +6,16 @@ export async function createTodo(input: Prisma.TodoCreateInput): Promise<Todo> {
   return await prisma.todo.create({ data: { ...input } });
 }
 
-export async function findAllByUserId(userId: number, orderBy: keyof Todo = 'createdAt'): Promise<Todo[]> {
+export async function findAllByUserId(userId: string | undefined, orderBy: keyof Todo = 'createdAt'): Promise<Todo[]> {
+  if (userId === undefined) {
+    throw new Error("User ID is undefined");
+  }
   return await prisma.todo.findMany({ where: { userId }, orderBy: { [orderBy]: 'asc' } });
 }
 
-export async function deleteTodoById(todoId: number, userId: number): Promise<Todo> {
+export async function deleteTodoById(todoId: number, userId: string | undefined): Promise<Todo> {
+  if (userId === undefined) {
+    throw new Error("User ID is undefined");
+  }
   return await prisma.todo.delete({ where: { id: todoId, userId } });
 }
